fix: refactored to allow deploy to industryUI
