+++ .storybook/.babelrc
@@ -1,14 +1,30 @@
 {
-  "env": {
-    "production": {
-      "presets": ["@babel/preset-env", "@babel/preset-react"]
-    },
-    "development": {
-      "presets": ["@babel/preset-env", "@babel/preset-react"],
-      "plugins": [
-        "@babel/plugin-proposal-object-rest-spread",
-        "@babel/plugin-proposal-class-properties"
-      ]
-    }
-  }
+  "presets": [
+    "next/babel",
+    "@babel/preset-env",
+    "@babel/preset-react"
+  ],
+  "plugins": [
+    "@babel/plugin-proposal-object-rest-spread",
+    "@babel/plugin-proposal-class-properties",
+    "@babel/plugin-transform-classes",
+    [
+      "styled-components", {
+        "displayName": true,
+        "ssr": true
+      }
+    ],
+    [
+      "module-resolver",
+      {
+        "root": [
+          "./"
+        ],
+        "alias": {
+          "api": "./api",
+          "config": "./config"
+        }
+      }
+    ]
+  ]
 }
