+++ .storybook/Container.js
@@ -1,10 +1,20 @@
 /**
  * Container
  */
+
+ // Style
+import { ThemeProvider } from 'styled-components'
+
+// UI
+import { Theme, ThemeStyle } from '../'
+
 export const Container = ({ story }) => {
   return (
-    <div style={{ margin: '1rem' }}>
-      {story()}
-    </div>
+    <ThemeProvider theme={Theme}>
+      <div style={{ margin: '1rem' }}>
+        {story()}
+      </div>
+      <ThemeStyle />
+    </ThemeProvider>
   )
 }
