--- .storybook/index.js
@@ -1,5 +1,0 @@
-/**
- * Components
- */
-export * from './Container'
-export * from './Theme'
