+++ .storybook/static/CNAME
@@ -1,0 +1,1 @@
+industryui.com
