+++ README.md
@@ -1,9 +1,9 @@
 # DryKISS: IndustryUI
 
 This is [DryKISS's](https://drykiss.com) react design system. It contains all of
-components used throughout our NextJS projects.
+our components used throughout our NextJS projects.
 
-It enables us to rapidly prototype any type of site and any scale.
+It enables us to rapidly prototype any type of site at any scale.
 
 ## Install
